Move OperatorClient to a separate directory
